sium

+++ uninstall.sh
@@ -15,7 +15,9 @@ systemctl daemon-reload
 rm -rf /opt/biscavolley/
 
 # Remove the environment file
-rm -rf /etc/biscavolley/
+if [[ ! " $@ " =~ " --preserve-etc " ]]; then
+    rm -rf /etc/biscavolley/
+fi
 
 # Notify the user about the user deletion
 echo "The user 'biscavolley' was not deleted automatically. Consider removing it manually, if needed."
